+++ modules/home/theme.nix
@@ -16,8 +16,10 @@
     gtk4.extraConfig.gtk-application-prefer-dark-theme = true;
   };
 
-  xdg.dataFile."icons/buuf-nestort".source =
-    "${pkgs.buuf-nestort-icon-theme}/share/icons/buuf-nestort";
+  xdg.dataFile."icons/buuf-nestort" = {
+    source = "${pkgs.buuf-nestort-icon-theme}/share/icons/buuf-nestort";
+    force = true;
+  };
 
   xdg.configFile."gtk-3.0/gtk.css".source =
     config.lib.file.mkOutOfStoreSymlink "${config.home.homeDirectory}/.cache/matugen/gtk.css";